pyproject.toml
[tool.poetry]
name = "pybtool"
version = "0.1.0"
description = "bluetooth tool"
authors = ["wwdeng <d102410241024@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "pybtool"}
]

[tool.poetry.dependencies]
python = "^3.8.10"
pyserial = "^3.5"
pyusb = "^1.2.1"
libusb = "^1.0.27.post1"
pyaudio = "^0.2.14"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
btool = "pybtool.main:main"

[tool.poetry.plugins."console_scripts"]
btool = "pybtool.main:main"
